requirements.txt
streamlit==1.32.0
langchain==0.1.9
chromadb==0.4.22
faiss-cpu==1.7.4
qdrant-client==1.7.0
python-dotenv==1.0.1
langchain-community==0.0.24
langchain-core==0.1.27
langchain-openai==0.0.3
pyyaml==6.0.1
python-multipart==0.0.9
typing-extensions>=4.11,<5.0
openai==1.55.3
tiktoken==0.5.2
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
async-timeout>=4.0.0 